Use the class instead.
