+++ doc/version-history.rst
@@ -5,7 +5,7 @@ Version History
 v0.1.1
 ======
 
-* Add the **utility_m2.py**.
+* Add the **diagnostics_m2.py**.
 
 v0.1.0
 ======
